agent/pyproject.toml
[project]
name = "pm-agent"
version = "0.1.0"
description = "Agent Service (FastAPI + LangGraph)"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "websockets>=12",
  "httpx>=0.27",
  "langchain-groq>=0.1",
  "langgraph>=0.2",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "typing-extensions>=4.9",
  "mem0ai>=0.1",
  "ollama>=0.2",
  "psycopg2-binary>=2.9.11",
  "pypdf>=4.0",
  "python-docx>=1.1",
  "minio>=7.2",
  "langchain-text-splitters>=0.2",
]

[tool.uv]
package = false
